Re-organize to finish earlier than 45 min.
