+++ src/mono/wasm/Wasm.Build.Tests/Blazor/BuildPublishTests.cs
@@ -143,60 +143,4 @@ public class BuildPublishTests : BlazorWasmTestBase
 
         Assert.Contains(razorClassLibraryFileName, lazyVal.EnumerateObject().Select(jp => jp.Name));
     }
-
-    [ConditionalTheory(typeof(BuildTestBase), nameof(IsUsingWorkloads))]
-    [InlineData("Debug")]
-    [InlineData("Release")]
-    public async Task BlazorBuildRunTest(string config)
-    {
-        string id = $"blazor_{config}_{Path.GetRandomFileName()}";
-        string projectFile = CreateWasmTemplateProject(id, "blazorwasm");
-
-        BlazorBuild(new BlazorBuildOptions(id, config, NativeFilesType.FromRuntimePack));
-        await BlazorRunForBuildWithDotnetRun(new BlazorRunOptions() { Config = config });
-    }
-
-    [ConditionalTheory(typeof(BuildTestBase), nameof(IsUsingWorkloads))]
-    [InlineData("Debug", false)]
-    [InlineData("Debug", true)]
-    [InlineData("Release", false)]
-    [InlineData("Release", true)]
-    public async Task BlazorPublishRunTest(string config, bool aot)
-    {
-        string id = $"blazor_{config}_{Path.GetRandomFileName()}";
-        string projectFile = CreateWasmTemplateProject(id, "blazorwasm");
-        if (aot)
-            AddItemsPropertiesToProject(projectFile, "<RunAOTCompilation>true</RunAOTCompilation>");
-
-        BlazorPublish(new BlazorBuildOptions(
-            id,
-            config,
-            aot ? NativeFilesType.AOT
-                : (config == "Release" ? NativeFilesType.Relinked : NativeFilesType.FromRuntimePack)));
-        await BlazorRunForPublishWithWebServer(new BlazorRunOptions() { Config = config });
-    }
-
-    private void BlazorAddRazorButton(string buttonText, string customCode, string methodName = "test", string razorPage = "Pages/Counter.razor")
-    {
-        string additionalCode = $$"""
-            <p role="{{methodName}}">Output: @outputText</p>
-            <button class="btn btn-primary" @onclick="{{methodName}}">{{buttonText}}</button>
-
-            @code {
-                private string outputText = string.Empty;
-                public void {{methodName}}()
-                {
-                    {{customCode}}
-                }
-            }
-        """;
-
-        // find blazor's Counter.razor
-        string counterRazorPath = Path.Combine(_projectDir!, razorPage);
-        if (!File.Exists(counterRazorPath))
-            throw new FileNotFoundException($"Could not find {counterRazorPath}");
-
-        string oldContent = File.ReadAllText(counterRazorPath);
-        File.WriteAllText(counterRazorPath, oldContent + additionalCode);
-    }
 }
